Adds a Vending machine and related specs
A vending machine is composed of a number of items along with a number
of coins. These coins can be helpful to tender change to a customer
who’d like to buy an item via a vending machine. The serial number
associated with a vending machine helps differentiate one from another.